requirements.txt
pygame-ce
json5
pytweening